requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
vertifi==2022.6.15
chardet==4.0.0
click==8.1.3
colorama==0.4.5
discord==1.7.3
discord.py==1.7.3
flask==2.1.2
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.1
multidict==6.0.2
numpy==1.22.2
pillow==9.1.1
replit==3.2.4
requests==2.28.0
typing-extensions==3.10.0.2
urllib3==1.26.9
werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0